+++ desktop/cmp/menu/impl/MenuItems.ts
@@ -76,7 +76,8 @@ export function parseMenuItems(
                 intent: item.intent,
                 className: item.className,
                 onClick: actionFn ? e => wait().then(() => actionFn(e, context)) : null, // do async to allow menu to close
-                disabled: item.disabled
+                disabled: item.disabled,
+                active: item.active
             };
 
             // Recursively parse any submenus